moving list_in_time_range to entity
