+++ backend/corpora/common/entities/project.py
@@ -1,9 +1,3 @@
-import typing
-from datetime import datetime
-
-import pytz
-from sqlalchemy import and_
-
 from .entity import Entity
 from ..corpora_orm import DbProject, DbProjectLink
 
@@ -16,17 +10,17 @@ class Project(Entity):
 
     @classmethod
     def create(
-        cls,
-        status: str,
-        name: str = "",
-        description: str = "",
-        owner: str = "",
-        s3_bucket: str = "",
-        tc_uri: str = "",
-        needs_attestation: bool = False,
-        processing_state: str = "",
-        validation_state: str = "",
-        links: list = None,
+            cls,
+            status: str,
+            name: str = "",
+            description: str = "",
+            owner: str = "",
+            s3_bucket: str = "",
+            tc_uri: str = "",
+            needs_attestation: bool = False,
+            processing_state: str = "",
+            validation_state: str = "",
+            links: list = None,
     ) -> "Project":
         """
         Need to check if one exists before creating
@@ -69,22 +63,3 @@ class Project(Entity):
         cls.db.session.add(new_db_object)
         cls.db.session.commit()
         return cls(new_db_object)
-
-    @classmethod
-    def list_in_time_range(cls, to_date: float = None, from_date: float = None) -> typing.List[typing.Dict]:
-        """
-
-        :param to_date: Filter dates earlier than this. Unix timestamp since the epoch in UTC timezone.
-        :param from_date: Filter dates later than this. Unix timestamp since the epoch in UTC timezone.
-        :return: The results in a list of dictionaries
-        """
-
-        filters = []
-        if to_date:
-            filters.append(DbProject.created_at <= datetime.fromtimestamp(to_date, tz=pytz.UTC))
-        if from_date:
-            filters.append(DbProject.created_at >= datetime.fromtimestamp(from_date, tz=pytz.UTC))
-
-        results = [result.to_dict() for result in cls.db.session.query(DbProject).filter(and_(*filters)).all()]
-
-        return results
